+++ packages/core/src/storage/types.ts
@@ -1,5 +1,5 @@
 import type { MetricResult, TestInfo } from '../eval';
-import { type MemoryConfig } from '../memory';
+import type { MemoryConfig } from '../memory/types';
 import { type WorkflowRunState } from '../workflows';
 
 export interface StorageColumn {
