+++ packages/core/src/telemetry/index.ts
@@ -1,4 +1,5 @@
 export * from './types';
 export * from './telemetry.decorators';
 export * from './utility';
-export { Telemetry, OTLPStorageExporter } from './telemetry';
+export { OTLPTraceExporter as OTLPStorageExporter } from './storage-exporter';
+export { Telemetry } from './telemetry';
